libdrm,intel: Factor source file lists into sources.mk
Factor the source file list for libdrm.so from Makefile.am into
sources.mk. Ditto for libdrm_intel.so.

This is in preparation for adding support for Android.  The sources.mk's
will be shared between autotools and Android.

Rationale: The most commonly changed parts of any makefile are the source
lists. So, by sharing the lists between the two build systems, we can
reduce the frequency at which modifications to the Linux build breaks the
Android build.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>
Signed-off-by: Sean V Kelley <sean.v.kelley@intel.com>

Conflicts:

	Makefile.am

Change-Id: If0cb137241dfa99f3ab608f127e2fd1f5c7886c2