REVERTME [VPG]: libdrm: Native sync user land
Modified DRM API to add support for native sync for batch buffer submissions.
The following functions are available in addition to the regular exec calls:

	drm_intel_bo_fence_exec
	drm_intel_bo_mrb_fence_exec
	drm_intel_bo_context_fence_exec

They optionally take a fence_in file descriptor to wait on before the request
starts and will accept a pointer to a fence_out file descriptor that the
kernel should fill in and return. It will be signalled when the batch buffer
completes or times-out.

The DRM_IOCTL_I915_GEM_EXECBUFFER2 has been changed to use DRM_IOWR so that
it can return data to user space.

Currently the KMD can only support completion fences so if the fence_in
parameter is valid then it will wait on the fence in user space *before*
calling the exec buffer IOCTL. This is a temporary measure as it blocks
the caller until the fence is signalled.

This is REVERTME because once the native sync is completely integrated with
GPU Scheduler these changes are not required.

For: GMIN-853
Tracked-On: https://jira01.devtools.intel.com/browse/GMIN-854
Change-Id: I7bd1a3121437fe6ca3b863135ce3e6a0068e41b8
Signed-off-by: ian-lister <ian.lister@intel.com>
Reviewed-on: http://android.intel.com:8080/127458
Reviewed-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Signed-off-by: Arun Siluvery <arun.siluvery@linux.intel.com>
Reviewed-on: https://android.intel.com/191711
Reviewed-by: Siluvery, Arun <arun.siluvery@intel.com>
Tested-by: Siluvery, Arun <arun.siluvery@intel.com>
Reviewed-by: Bloomfield, Jon <jon.bloomfield@intel.com>