REVERTME [VPG]: Added dummy DISP_SCREEN_CONTROL ioctl to satisfy hwc
The current VPG hardware composer implementation has an unused function which
calls the now obsolete DISPL_SCREEN_CONTROL ioctl. This ioctl has no
corresponding implementation in the kernel.

Until the VPG source can be fixed this patch defines the references symbols
to enable hwc to build correctly.

This patch should be removed as soon as hwc is fixed

Change-Id: Ifd5e99d9d683691d6cc87852de83a4b39b8e2ba2
Signed-off-by: Jon Bloomfield <jon.bloomfield@intel.com>