REVERTME [VPG]: Zero DR4 when num_cliprects==0
This patch works around an issue in the VPG media driver.

All callers of execbuffer are now required to 0 unused
arguments.

The kernel now checks that DR1, DR4 and cliprects_ptr are all 0 if
num_cliprects is 0 and errors if they are not.

Currently, the media driver is not zeroing DR4
despite not supplying any cliprects.

REVERTME: This is not a permanent fix because it is expected that the
execbuffer interface may allow these fields to be reused for other
purposes in the future, so we should really allow them to pass through
normally, and fix this in the media driver instead.

I've tagged the Jira below with For: instead of Issue: so that the Jira
is not closed until the proper patch is incorporated into the userland
media driver, and this patch reverted.

For: GMIN-1837
Change-Id: I7428e20a23831c8f1c9d0c46346b2bd68a975091
Signed-off-by: Jon Bloomfield
Reviewed-on: https://android.intel.com/209473
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Bloomfield, Jon <jon.bloomfield@intel.com>
Tested-by: Bloomfield, Jon <jon.bloomfield@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Harris, Chris <chris.harris@intel.com>
Tested-by: Harris, Chris <chris.harris@intel.com>
Reviewed-by: Siluvery, Arun <arun.siluvery@intel.com>
Reviewed-by: Webb, Russell <russell.webb@intel.com>