SQUASHME! [VPG]: libdrm: Updates the dpst interface with scenario based enabling.
This is a continuation to patch that adds dpst libdrm interface function.
The extended ioctl scheme is enabled much later than the first patch for
dpst interface. A property which enables dpst only on video playback is added
to the dpst initialize ioctl by this time. This patch updates the interface with
this new property.

SQUASHME! - This patch should be squashed into the following existing patch:

        Author: Minu Mathai <minu.mathai@intel.com>
        Date:   Dec 12, 2014 3:25 PM
        Change-Id: Ic495a84b5b8f2958978e6891ee19c605fe35752f
        MUST_REBASE [VPG]: libdrm: add interface function for I915_dpst

Issue: GMINL-7162
Change-Id: I70cdaeb2b1414d9ec148fb8df6214dda92d9f90e
Signed-off-by: Minu Mathai <minu.mathai@intel.com>
Reviewed-on: https://android.intel.com/337603
Reviewed-by: Yi, Feitong <feitong.yi@intel.com>
Tested-by: Yi, Feitong <feitong.yi@intel.com>